adding option to delete local copy of the file

+++ hsfiles_jupyter/delete_file.py
@@ -9,8 +9,11 @@ from .utils import (
 )
 
 
-async def delete_file_from_hydroshare(file_path: str):
-    """Deletes a file 'file_path' from HydroShare resource as well as from the local filesystem."""
+async def delete_file_from_hydroshare(file_path: str, delete_local_file: bool = False):
+    """
+    Deletes a file 'file_path' from HydroShare resource as well as from the
+    local filesystem if delete_local_file is set to True.
+    """
 
     rfc_manager = ResourceFileCacheManager()
     try:
@@ -37,7 +40,6 @@ async def delete_file_from_hydroshare(file_path: str):
         err_msg = f"File {res_info.hs_file_path} doesn't exist in HydroShare resource: {res_info.resource_id}"
         return {"error": err_msg}
 
-    local_file_to_delete_full_path = get_local_absolute_file_path(file_path)
     try:
         # deleting from HydroShare
         res_info.resource.file_delete(hs_file_to_delete)
@@ -49,13 +51,19 @@ async def delete_file_from_hydroshare(file_path: str):
                    f' resource: {res_info.resource_id}. Error: {hs_error}')
         logger.error(err_msg)
         return {"error": err_msg}
-    try:
-        # deleting from local filesystem
-        os.remove(local_file_to_delete_full_path)
-        return {"success": f"File {res_info.hs_file_path} was deleted from HydroShare resource: {res_info.resource_id}"}
-    except Exception as e:
-        os_error = str(e)
-        err_msg = (f'File {res_info.hs_file_path} was deleted from HydroShare resource: {res_info.resource_id}\n.'
-                   f' NOTE: However, the local file could not be deleted. Error: {os_error}')
-        logger.error(err_msg)
-        return {"success": err_msg}
+
+    delete_success_msg = f'File {res_info.hs_file_path} was deleted from HydroShare resource: {res_info.resource_id}'
+    if delete_local_file:
+        local_file_to_delete_full_path = get_local_absolute_file_path(file_path)
+        try:
+            # deleting from local filesystem
+            os.remove(local_file_to_delete_full_path)
+            return {"success": delete_success_msg}
+        except Exception as e:
+            os_error = str(e)
+            err_msg = (f'File {res_info.hs_file_path} was deleted from HydroShare resource: {res_info.resource_id}\n.'
+                       f' NOTE: However, the local file could not be deleted. Error: {os_error}')
+            logger.error(err_msg)
+            return {"success": err_msg}
+    else:
+        return {"success": delete_success_msg}
